dts: sdm845: import DT from Linux
With driver support improved, replace the u-boot specific devicetrees
for SDM845 targets with the Linux DT for the same. The devicetree and
include files are pulled in directly from Linux 6.5-rc5 and should be
kept in-sync as needed, not modified.

Adjust the dragonboard845c and starqltechn DTS files to support these
changes.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>